+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Improvements
 
 - feat(iam): Automatically setup OIDC provider
+  [#281](https://github.com/pulumi/pulumi-eks/pull/281)
 
 ## 0.18.16 (Released November 7, 2019)