+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### Improvements
 
-- feat(nodes): add createManagedNodeGroup(), and ManagedNodeGroup resource
+## 0.18.17 (Released December 3, 2019)
+
+### Improvements
+
+- feat(nodes): add createManagedNodeGroup
   [#280](https://github.com/pulumi/pulumi-eks/pull/280)
 
 ## 0.18.16 (Released November 7, 2019)